fixed up my small mistake
I did'nt add the title tag to the header div for some reason; I forgot to do so.

I also fixed up my CSS and removed the white border and made the text easier to read.

Finally, I moved the legend back to the left; I am not sure why it was centered in the first place.